Create a separate ConversionUtilPrivilegeTest
